+++ Robust.Shared/GameObjects/EntityManager.Components.cs
@@ -109,7 +109,7 @@ namespace Robust.Shared.GameObjects
             DebugTools.AssertOwner(uid, metadata);
             metadata ??= GetComponent<MetaDataComponent>(uid);
             DebugTools.Assert(metadata.EntityLifeStage == EntityLifeStage.PreInit);
-            metadata.EntityLifeStage = EntityLifeStage.Initializing;
+            SetLifeStage(metadata, EntityLifeStage.Initializing);
 
             // Initialize() can modify the collection of components. Copy them.
             FixedArray32<IComponent?> compsFixed = default;
@@ -136,7 +136,7 @@ namespace Robust.Shared.GameObjects
 
 #endif
             DebugTools.Assert(metadata.EntityLifeStage == EntityLifeStage.Initializing);
-            metadata.EntityLifeStage = EntityLifeStage.Initialized;
+            SetLifeStage(metadata, EntityLifeStage.Initialized);
         }
 
         public void StartComponents(EntityUid uid)